build(MRT): :fire: change repo structure
change repo structure(move some file to utils)
