+++ airflow/utils/etl/mrt_realtime_arrival.py
@@ -9,7 +9,6 @@ import numpy as np
 from sqlalchemy import create_engine, exc
 from zoneinfo import ZoneInfo
 from google.cloud import storage
-from _dev.MRT_ETL_function.upload_to_gcs_function import upload_to_bucket_string
 load_dotenv()
