+++ airflow/utils/etl/mrt_realtime_crowded_BR.py
@@ -9,7 +9,6 @@ import numpy as np
 from sqlalchemy import create_engine, exc
 from zoneinfo import ZoneInfo
 from google.cloud import storage
-from MRT_ETL_function.upload_to_gcs_function import upload_to_bucket_string
 load_dotenv()
