+++ airflow/utils/gcp/create_external_table.py
@@ -2,7 +2,6 @@ import os
 import pandas as pd
 from google.cloud import bigquery
 from google.oauth2.service_account import Credentials
-import db_dtypes
 
 
 def SRC_youbike_gcs_to_bq_before0504(dataset_name: str, create_table_name: str, client: bigquery.Client) -> None:
